requirements.txt
streamlit>=1.23,<2
matplotlib==3.8.4
pandas==2.1.4
wordcloud==1.9.4
google-api-python-client==2.185.0
python-dotenv==1.1.1
praw==7.8.1
tensorflow